+++ cross/auto_parameters/shared/evaluate_model/evaluate_model.py
@@ -1,7 +1,7 @@
 import numpy as np
 from sklearn.compose import ColumnTransformer, make_column_selector
 from sklearn.impute import SimpleImputer
-from sklearn.model_selection import KFold, cross_val_score
+from sklearn.model_selection import cross_val_score
 from sklearn.pipeline import Pipeline
 
 
@@ -32,13 +32,13 @@ def evaluate_model(
     y,
     model,
     scoring,
+    cv=5,
+    groups=None,
     transformer=None,
 ):
-    # Build pipeline with optional transformer and model
     pipe = build_pipeline(model, transformer)
-
-    # Perform cross-validation
-    cv = KFold(n_splits=5, shuffle=True, random_state=42)
-    scores = cross_val_score(pipe, x, y, scoring=scoring, cv=cv, n_jobs=-1)
+    scores = cross_val_score(
+        pipe, x, y, scoring=scoring, cv=cv, groups=groups, n_jobs=-1
+    )
 
     return np.mean(scores)